refactor: abstract counter
